requirements.txt
accelerate
bitsandbytes
diffusers
huggingface-hub
numpy
pandas
peft
pillow
safetensors
scikit-learn
scipy
sentence-transformers
tensorboard
tokenizers
torch
torchaudio
torchvision
tqdm 
transformers
supervision
inference-gpu[yolo-world]==0.9.12rc1
supervision==0.19.0rc3
